extract credential-decision logic into config/credentials.ts
